Move some Op in new _memoryoperations.py and rename _internalobjects.py in _commands.py
The file _operations.py with all the docstring was becoming too big, so i
have splitted the file putting the operation which touch the memory into
an appropriate container module _memoryoperations.py .
The file _internalobjects.py was renamed in _commands.py for more consistency.
